+++ src/Form/BulkImportForm.php
@@ -20,9 +20,9 @@ final class BulkImportForm extends FormBase {
   /**
    * Constructs the bulk import form. */
   public function __construct(
-    private readonly EntityTypeManagerInterface $entityTypeManager,
-    private readonly CollectionDocumentManager $manager,
-    private readonly AccountProxyInterface $currentUser,
+    protected EntityTypeManagerInterface $entityTypeManager,
+    protected CollectionDocumentManager $manager,
+    protected AccountProxyInterface $currentUser,
   ) {}
 
   /**